#17 rename : 카테고리 파일명 및 경로 변경
