+++ src/app/(contents)/main/card/CommonCard.tsx
@@ -1,4 +1,4 @@
-import S from './Card.module.scss';
+import S from './CommonCard.module.scss';
 
 export default function CommonCard({ children }: { children: React.ReactNode }) {
   return <div className={`${S.card} ${S.cardLarge}`}>{children}</div>;
