--- src/app/(contents)/main/category/category.module.scss
@@ -1,19 +1,0 @@
-@use '@/styles/abstracts/_core.scss' as *;
-
-.categoryContainer {
-  display: flex;
-  gap: 24px;
-}
-
-.category {
-  width: 127px;
-  height: 58px;
-
-  text-align: center;
-  padding: 16px;
-  border-radius: 15px;
-  border: 1px solid $green300;
-  background-color: $white;
-  @include text2LgMedium;
-  color: $green300;
-}
